Remove AuthService & Database abstract classes
