+++ lib/app/home/jobs/jobs_page.dart
@@ -14,7 +14,7 @@ import 'package:starter_architecture_flutter_firebase/services/database.dart';
 class JobsPage extends StatelessWidget {
   Future<void> _delete(BuildContext context, Job job) async {
     try {
-      final database = Provider.of<Database>(context, listen: false);
+      final database = Provider.of<FirestoreDatabase>(context, listen: false);
       await database.deleteJob(job);
     } on PlatformException catch (e) {
       PlatformExceptionAlertDialog(
@@ -41,7 +41,7 @@ class JobsPage extends StatelessWidget {
   }
 
   Widget _buildContents(BuildContext context) {
-    final database = Provider.of<Database>(context, listen: false);
+    final database = Provider.of<FirestoreDatabase>(context, listen: false);
     return StreamBuilder<List<Job>>(
       stream: database.jobsStream(),
       builder: (context, snapshot) {
