+++ lib/services/database.dart
@@ -6,32 +6,19 @@ import 'package:starter_architecture_flutter_firebase/app/home/models/job.dart';
 import 'package:starter_architecture_flutter_firebase/services/api_path.dart';
 import 'package:starter_architecture_flutter_firebase/services/firestore_service.dart';
 
-abstract class Database {
-  Future<void> setJob(Job job);
-  Future<void> deleteJob(Job job);
-  Stream<List<Job>> jobsStream();
-  Stream<Job> jobStream({@required String jobId});
-
-  Future<void> setEntry(Entry entry);
-  Future<void> deleteEntry(Entry entry);
-  Stream<List<Entry>> entriesStream({Job job});
-}
-
 String documentIdFromCurrentDate() => DateTime.now().toIso8601String();
 
-class FirestoreDatabase implements Database {
+class FirestoreDatabase {
   FirestoreDatabase({@required this.uid}) : assert(uid != null);
   final String uid;
 
   final _service = FirestoreService.instance;
 
-  @override
   Future<void> setJob(Job job) async => await _service.setData(
         path: APIPath.job(uid, job.id),
         data: job.toMap(),
       );
 
-  @override
   Future<void> deleteJob(Job job) async {
     // delete where entry.jobId == job.jobId
     final allEntries = await entriesStream(job: job).first;
@@ -44,29 +31,24 @@ class FirestoreDatabase implements Database {
     await _service.deleteData(path: APIPath.job(uid, job.id));
   }
 
-  @override
   Stream<Job> jobStream({@required String jobId}) => _service.documentStream(
         path: APIPath.job(uid, jobId),
         builder: (data, documentId) => Job.fromMap(data, documentId),
       );
 
-  @override
   Stream<List<Job>> jobsStream() => _service.collectionStream(
         path: APIPath.jobs(uid),
         builder: (data, documentId) => Job.fromMap(data, documentId),
       );
 
-  @override
   Future<void> setEntry(Entry entry) async => await _service.setData(
         path: APIPath.entry(uid, entry.id),
         data: entry.toMap(),
       );
 
-  @override
   Future<void> deleteEntry(Entry entry) async =>
       await _service.deleteData(path: APIPath.entry(uid, entry.id));
 
-  @override
   Stream<List<Entry>> entriesStream({Job job}) =>
       _service.collectionStream<Entry>(
         path: APIPath.entries(uid),
