+++ lib/services/firebase_auth_service.dart
@@ -1,7 +1,22 @@
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:starter_architecture_flutter_firebase/services/auth_service.dart';
+import 'package:flutter/foundation.dart';
 
-class FirebaseAuthService implements AuthService {
+@immutable
+class User {
+  const User({
+    @required this.uid,
+    this.email,
+    this.photoUrl,
+    this.displayName,
+  });
+
+  final String uid;
+  final String email;
+  final String photoUrl;
+  final String displayName;
+}
+
+class FirebaseAuthService {
   final FirebaseAuth _firebaseAuth = FirebaseAuth.instance;
 
   User _userFromFirebase(FirebaseUser user) {
@@ -16,18 +31,15 @@ class FirebaseAuthService implements AuthService {
     );
   }
 
-  @override
   Stream<User> get onAuthStateChanged {
     return _firebaseAuth.onAuthStateChanged.map(_userFromFirebase);
   }
 
-  @override
   Future<User> signInAnonymously() async {
     final AuthResult authResult = await _firebaseAuth.signInAnonymously();
     return _userFromFirebase(authResult.user);
   }
 
-  @override
   Future<User> signInWithEmailAndPassword(String email, String password) async {
     final AuthResult authResult = await _firebaseAuth
         .signInWithCredential(EmailAuthProvider.getCredential(
@@ -37,7 +49,6 @@ class FirebaseAuthService implements AuthService {
     return _userFromFirebase(authResult.user);
   }
 
-  @override
   Future<User> createUserWithEmailAndPassword(
       String email, String password) async {
     final AuthResult authResult = await _firebaseAuth
@@ -45,22 +56,16 @@ class FirebaseAuthService implements AuthService {
     return _userFromFirebase(authResult.user);
   }
 
-  @override
   Future<void> sendPasswordResetEmail(String email) async {
     await _firebaseAuth.sendPasswordResetEmail(email: email);
   }
 
-  @override
   Future<User> currentUser() async {
     final FirebaseUser user = await _firebaseAuth.currentUser();
     return _userFromFirebase(user);
   }
 
-  @override
   Future<void> signOut() async {
     return _firebaseAuth.signOut();
   }
-
-  @override
-  void dispose() {}
 }
