+++ test/mocks.dart
@@ -1,8 +1,8 @@
-import 'package:starter_architecture_flutter_firebase/services/auth_service.dart';
 import 'package:flutter/material.dart';
 import 'package:mockito/mockito.dart';
+import 'package:starter_architecture_flutter_firebase/services/firebase_auth_service.dart';
 
-class MockAuthService extends Mock implements AuthService {}
+class MockAuthService extends Mock implements FirebaseAuthService {}
 
 class MockWidgetsBinding extends Mock implements WidgetsBinding {}
