+++ test/sign_in_page_test.dart
@@ -1,12 +1,12 @@
 import 'dart:async';
 
 import 'package:starter_architecture_flutter_firebase/app/sign_in/sign_in_page.dart';
-import 'package:starter_architecture_flutter_firebase/routing/router.gr.dart';
-import 'package:starter_architecture_flutter_firebase/services/auth_service.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:provider/provider.dart';
+import 'package:starter_architecture_flutter_firebase/routing/router.gr.dart';
+import 'package:starter_architecture_flutter_firebase/services/firebase_auth_service.dart';
 
 import 'mocks.dart';
 
@@ -29,7 +29,7 @@ void main() {
     await tester.pumpWidget(
       MultiProvider(
         providers: [
-          Provider<AuthService>(
+          Provider<FirebaseAuthService>(
             create: (_) => mockAuthService,
           ),
         ],
